Emitting from JSON is now working, tweaked emitIcon
